Check guild before message delete

+++ src/MessageHandler.ts
@@ -51,6 +51,7 @@ export class MessageHandler {
     if (joinDate && currentDate.getTime() - joinDate.getTime() < 604800000) {
         if (str_message.includes('http://') || (str_message.includes('https://') && !permittedWebsites.some((website) => str_message.includes(website)))) {
             if (!guildMember?.roles.cache.some((role) => Object.values(roles).includes(role.name))) {
+              if(message.guildId === '852104066401173514')
                 await message.delete();
             }
         }
